New fit for rhoV
ipole's fits have traditionally assumed that Theta_e >> 1,
i.e. that the electrons involved are hot.  For nearly all
emission and absorption, this is the case.  However, Faraday
rotation can occur much farther out in the disk, where temperatures
fall to Theta_e < 1, and sometimes Theta_e < 0.1.

This commit adds two special cases to the calculation of the rotativity
of the linear polarization vector, rho_V, for low temperature.

For Theta_e < 3, it switches from fast high-temperature approximations of
the Bessel functions to more accurate versions from gsl_sf_bessel
For Theta_e < 0.01 it uses the constant low-temperature limit of the
rotation measure, identical to e.g. interstellar medium calculations.

This change does not dramatically affect EVPA maps weighted by emission
in blurred images of M87, as the system is viewed near the pole and sight lines
do not travel through cold material.  For lower-inclination systems
the change is O(1).  To the extent the gas state in this outer region is unphysical,
we may want to implement a manual cut to prevent EVPA "scrambling" due to very high
Faraday depth.