Tentative fix for very high-absorbtivity regions
This commit adds a high-absorptivity limit expression to the analytic
solutions ipole uses when evolving Stokes parameters in fluid frame.

Previously, ipole had a low-absorptivity limit, but not a limit for
(equally volatile) large absorptivities, which have begun to crop
up in rare cases.

This fix appears to be relevant only in new GRMHD runs/dumps, and
then only in the midplane for large Rhigh.

This also cleans up some debugging prints likely to be irrelevant.