Correct a very old typo
One of the coefficients in the j_V fit taken from Dexter (2016) was mistyped in the last two digits, as 1.81348 instead of 1.81384.  The typo appears to date back to the very first version of ipole.  Correcting this should have no discernible effect on transport whatsoever, but might make the code easier to read & compare in the future.